// src/API.ts
export type Note = {
  __typename: "Note",
  id: string,
  name: string,
  description?: string | null,
  image?: string | null,
  createdAt: string,
  updatedAt: string,
};